+++ client/client.go
@@ -2,15 +2,13 @@ package client
 
 import (
 	"context"
-	"github.com/libp2p/go-libp2p-core/host"
 	"github.com/rovergulf/rbn/pkg/traceutil"
-	"github.com/rovergulf/rbn/proto"
+	"github.com/rovergulf/rbn/rpc"
 	"go.uber.org/zap"
 )
 
 // NetherClient represents Rovergulf BlockChain Network gRPC client interface
 type NetherClient struct {
-	host   host.Host
 	logger *zap.SugaredLogger
 	tracer *traceutil.Tracer
 }
@@ -19,7 +17,6 @@ func NewClient(ctx context.Context, lg *zap.SugaredLogger, addr string) (*Nether
 
 	return &NetherClient{
 		logger: lg,
-		host:   nil,
 	}, nil
 }
 
@@ -29,12 +26,10 @@ func (c *NetherClient) HealthCheck(ctx context.Context) error {
 }
 
 func (c *NetherClient) Stop() {
-	if err := c.host.Close(); err != nil {
-		c.logger.Errorf("Unable to close p2p conn: %s", err)
-	}
+	// TBD
 }
 
-func (c *NetherClient) MakeCall(ctx context.Context, cmd proto.Command, ent proto.Entity, req []byte) (*proto.CallResponse, error) {
+func (c *NetherClient) MakeCall(ctx context.Context, req *rpc.ApiRequest) (*rpc.ApiResponse, error) {
 
 	return nil, nil
 }