+++ commands/tx_pool.go
@@ -1,0 +1,9 @@
+package commands
+
+import "github.com/rovergulf/rbn/core/types"
+
+// TBD or moved
+type TxPool struct {
+	pending types.Transactions
+	queued  types.Transactions
+}