+++ node/http_utils.go
@@ -9,6 +9,7 @@ import (
 	"github.com/prometheus/prom2json"
 	"github.com/rovergulf/rbn/params"
 	"github.com/rovergulf/rbn/pkg/resutil"
+	"github.com/spf13/viper"
 	"go.uber.org/zap"
 	"net/http"
 	"strings"
@@ -107,27 +108,22 @@ func (n *Node) WalkRoutes(w http.ResponseWriter, r *http.Request) {
 
 		pathTemplate, err := route.GetPathTemplate()
 		if err == nil {
-			//h.Logger.Debug("ROUTE: ", pathTemplate)
 			res["route"] = pathTemplate
 		}
 		pathRegexp, err := route.GetPathRegexp()
 		if err == nil {
-			//h.Logger.Debug("Path regexp: ", pathRegexp)
 			res["regexp"] = pathRegexp
 		}
 		queriesTemplates, err := route.GetQueriesTemplates()
 		if err == nil {
-			//h.Logger.Debug("Queries templates: ", strings.Join(queriesTemplates, ","))
 			res["queries_templates"] = strings.Join(queriesTemplates, ",")
 		}
 		queriesRegexps, err := route.GetQueriesRegexp()
 		if err == nil {
-			//h.Logger.Debug("Queries regexps: ", strings.Join(queriesRegexps, ","))
 			res["queries_regexps"] = strings.Join(queriesRegexps, ",")
 		}
 		methods, err := route.GetMethods()
 		if err == nil {
-			//h.Logger.Debug("Methods: ", strings.Join(methods, ","))
 			res["methods"] = methods
 		}
 
@@ -136,19 +132,11 @@ func (n *Node) WalkRoutes(w http.ResponseWriter, r *http.Request) {
 	})
 	if err != nil {
 		n.logger.Error(err)
+		n.httpResponse(w, err.Error(), http.StatusInternalServerError)
+		return
 	}
 
 	n.httpResponse(w, results)
-}
-
-func (n *Node) healthCheck(w http.ResponseWriter, r *http.Request) {
-	n.httpResponse(w, map[string]interface{}{
-		"http_status": http.StatusOK,
-		"timestamp":   time.Now().Unix(),
-		"run_date":    params.RunDate.Format(time.RFC1123),
-		"node_status": "healthy",
-		"in_gamble":   n.inGenRace,
-	})
 }
 
 func (n *Node) DiscoverMetrics(w http.ResponseWriter, r *http.Request) {
@@ -156,11 +144,10 @@ func (n *Node) DiscoverMetrics(w http.ResponseWriter, r *http.Request) {
 	ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
 	defer cancel()
 
-	metricsUrl := fmt.Sprintf("%s/metrics", n.metadata.HttpApiAddress())
+	metricsUrl := fmt.Sprintf("%s/metrics", n.HttpApiAddress())
 	req, err := http.Get(metricsUrl)
 	if err != nil {
 		n.logger.Errorf("Unable to send request to prometheus metrics: %s", err)
-		w.WriteHeader(http.StatusInternalServerError)
 		n.httpResponse(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
@@ -180,4 +167,21 @@ func (n *Node) DiscoverMetrics(w http.ResponseWriter, r *http.Request) {
 	}
 
 	n.httpResponse(w, result)
+}
+
+func (n *Node) healthCheck(w http.ResponseWriter, r *http.Request) {
+	n.httpResponse(w, map[string]interface{}{
+		"version":     "v" + params.MetaVersion,
+		"http_status": http.StatusOK,
+		"timestamp":   time.Now().Unix(),
+		"run_date":    params.RunDate.Format(time.RFC1123),
+		"healthy":     true,
+		"http":        fmt.Sprintf("%s:%s", viper.GetString("http.addr"), viper.GetString("http.port")),
+		"p2p":         fmt.Sprintf("%s:%s", viper.GetString("node.addr"), viper.GetString("node.port")),
+	})
+}
+
+// HttpApiAddress returns full API server URL
+func (n *Node) HttpApiAddress() string {
+	return fmt.Sprintf("%s://%s", viper.GetString("http.addr"), viper.GetString("http.port"))
 }