+++ node/node_db.go
@@ -1,0 +1,5 @@
+package node
+
+type DB struct {
+	DataDir string `json:"data_dir" yaml:"data_dir"`
+}