+++ CHANGELOG.md
@@ -4,6 +4,44 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased] v0.1.0
 
+## 9 Sep 2021
+
+### Added
+
+### Changed
+
+### Fixed
+
+### Removed
+
+
+## 8 Sep 2021
+
+### Added
+- GitHub actions Go fmt and vet checks (tests would be added as better design appears)
+- new p2p handling templates based on Etherium `p2p` 
+
+### Changed
+
+### Fixed
+
+### Removed
+- gRPC and protobuf usage
+
+
+## 6 Sep 2021
+
+### Changed
+- `proto` package renamed and moved to `node/pb`. `scripts/genproto.sh` fixed as well
+
+### Fixed
+- Downgrade go-multiaddr and libp2p-go-core libs, due runtime error
+
+### Removed
+- libp2p usage, there is etherium library, simply working right up here. What did I even tried to do?
+- legacy `etcd/raft` import
+
+
 ## 5 Sep 2021
 
 ### Added
@@ -12,6 +50,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 ### Fixed
+- close topic, dht and subs on graceful shutdown
 
 ### Removed